chore: 更新构建环境和README
- 更改README为中文
- 升级构建环境引入webpack
