+++ index.html
@@ -7,7 +7,7 @@
   <meta name="description" content="myphotogallery">
   <title>MyPhotoGallery-看电影也可以很有趣</title>
   <link rel="shortcut icon" href="./favicon.png" type="image/x-icon">
-  <link rel="stylesheet" type="text/css" href="./assets/src/index.css">
+  <link rel="stylesheet" type="text/css" href="./src/index.css">
 </head>
 <body class="noscroll">
   <div id="app">
@@ -121,6 +121,6 @@
     <div>&copy; shenlvmeng 2017</div>
   </footer>
   <script src="./dist/vue.min.js"></script>
-  <script src="./assets/src/index.js"></script>
+  <script src="./dist/index.js"></script>
 </body>
 </html>
