+++ package.json
@@ -24,13 +24,14 @@
   "homepage": "https://github.com/shenlvmeng/MyPhotoGallery#readme",
   "devDependencies": {
     "babel-core": "^6.26.0",
-    "babel-preset-es2015": "^6.24.1",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.1",
     "gulp": "^3.9.1",
-    "gulp-babel": "^7.0.0",
     "gulp-clean-css": "^3.0.0",
     "gulp-jsonminify": "^1.0.0",
     "gulp-rename": "^1.2.2",
     "gulp-uglify": "^3.0.0",
-    "gulp-webserver": "^0.9.1"
+    "gulp-webserver": "^0.9.1",
+    "webpack-stream": "^4.0.0"
   }
 }
